both left and right pane updated
This update is still under progress in the branch. The most important thing to notice here is how events from the database will be displayed on the webpage. At the moment I'm using a Java script array that temporarily stores demo events data - which then I display on the right pane. Check out the Branch, (do not merge with main yet),  After checking out, navigate to march and April 2024 and see what happens on the right pane. This is the mode or style or design that display of events will take. We can of course adjust the css and other small components like icon sizes later, but key thing here is for you to have an idea of how we intend to display events on the webpage to avoid any potential incompatibility